Finally zoom capable
Had to rewrite a bit of the old Model code for this to work. Can now
zoom on a fractal infinitely. Still needs visual box to display the
zoomed area. Also… because I rewrote the model code so much… I broke
the unit tests… those need to be fixed too :/